refactor: rename sources module to movies
