+++ mwc/movies/cli.py
